+++ lib/src/main/java/org/asamk/signal/manager/api/Message.java
@@ -7,6 +7,7 @@ public record Message(
         String messageText,
         List<String> attachments,
         List<AttachmentDimensions> attachmentDimensions,
+        List<String> attachmentBlurHashes,
         boolean viewOnce,
         boolean voiceNote,
         List<Mention> mentions,
@@ -18,16 +19,7 @@ public record Message(
         boolean urgent
 ) {
 
-    public record AttachmentDimensions(int width, int height) {
-
-        public static final AttachmentDimensions UNKNOWN = new AttachmentDimensions(0, 0);
-
-        public AttachmentDimensions {
-            if (width < 0 || height < 0 || (width == 0) != (height == 0)) {
-                throw new IllegalArgumentException("Dimensions must both be positive or both zero");
-            }
-        }
-    }
+    public record AttachmentDimensions(int width, int height) {}
 
     public record Mention(RecipientIdentifier.Single recipient, int start, int length) {}