+++ src/main/java/org/asamk/signal/dbus/DbusSignalImpl.java
@@ -243,6 +243,7 @@ public class DbusSignalImpl implements Signal, AutoCloseable {
             final var message = new Message(messageText,
                     attachments,
                     List.of(),
+                    List.of(),
                     false,
                     false,
                     List.of(),
@@ -411,6 +412,7 @@ public class DbusSignalImpl implements Signal, AutoCloseable {
             final var message = new Message(messageText,
                     attachments,
                     List.of(),
+                    List.of(),
                     false,
                     false,
                     List.of(),
@@ -458,6 +460,7 @@ public class DbusSignalImpl implements Signal, AutoCloseable {
         try {
             final var message = new Message(messageText,
                     attachments,
+                    List.of(),
                     List.of(),
                     false,
                     false,