+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `send --attachment-dimensions` to supply displayed attachment dimensions for correctly sized video placeholders
+- `send --attachment-dimensions` and `--attachment-blurhash` to set the placeholder shown before an attachment is downloaded
 
 ## [0.14.8] - 2026-09-10